infra_deploy/app/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
pandas==2.2.3
boto3==1.37.11
botocore==1.37.11
s3transfer==0.11.2
mlflow==2.20.4
python-dotenv==1.0.1
tqdm==4.67.1
catboost==1.2.7
numpy==1.26.4
urllib3==1.26.5
loguru==0.7.3
starlette_exporter==0.23.0
prometheus-client==0.21.1
